src/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coloredlogs==10.0
Flask==1.1.1
Flask-BasicAuth==0.2.0
gunicorn==19.9.0
humanfriendly==4.18
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
python-dotenv==0.10.3
requests==2.22.0
urllib3==1.25.6
Werkzeug==0.16.0
